Modernizes e2e session_persistence test re: NewBaseTestingSuite.
Uses base test suite for session_persistence e2e test.

The flake was
https://github.com/kgateway-dev/kgateway/actions/runs/17080668371/job/48433512049?pr=12072#step:6:110

'exit status 7' for curl suggests that the gateway is the issue, not the echo
Service. This is strongly suggested by running `kubectl scale --replicas=0 -n
default deploy/gw-cookie` and then running `kubectl exec --container=curl curl
-n curl -- curl --connect-timeout 1 --max-time 5 -v -s --connect-timeout 3
--max-time 3 -H 'Host: echo.local' -H 'Cookie: '
http://gw-cookie.default.svc.cluster.local:8080/`

This particular flakiness should be gone, perhaps thanks to the contemporary
and already landed #12233, or perhaps thanks to this change.

Signed-off-by: David L. Chandler <david.chandler@solo.io>