tests: test/deployer golden files are more end-to-end
Before, they were helm's output. Soon we'll introduce a strategic-merge-patch
step after that, so let's make the #13018 code review easier by switching to
golden files that are marshalled YAML of the objects we'll deploy (the identity
function (for now) on helm's output but with alphabetical map keys since helm
isn't producing the golden files).

Signed-off-by: David L. Chandler <david.chandler@solo.io>